fix loading logo for event, decompose event item screen, implement invites
