+++ app/src/main/java/com/distributedLab/rarime/api/points/PointsAPI.kt
@@ -33,7 +33,8 @@ interface PointsAPI {
     @GET("${BaseConfig.RELAYER_URL}/integrations/rarime-points-svc/v1/public/balances/{nullifier}")
     suspend fun getPointsBalance(
         @Path("nullifier") nullifier: String,
-        @Header("Authorization") authorization: String
+        @Header("Authorization") authorization: String,
+        @QueryMap queryParams: Map<String, String>
     ): PointsBalanceBody
 
     @POST("${BaseConfig.RELAYER_URL}/integrations/rarime-points-svc/v1/public/balances/{nullifier}/verifypassport")
