+++ app/src/main/java/com/distributedLab/rarime/modules/main/MainScreen.kt
@@ -1,7 +1,6 @@
 package com.distributedLab.rarime.modules.main
 
 import android.annotation.SuppressLint
-import android.util.Log
 import androidx.compose.animation.ExitTransition
 import androidx.compose.animation.core.RepeatMode
 import androidx.compose.animation.core.animateFloatAsState
@@ -11,7 +10,6 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
@@ -55,7 +53,7 @@ import com.distributedLab.rarime.modules.profile.ProfileScreen
 import com.distributedLab.rarime.modules.profile.ThemeScreen
 import com.distributedLab.rarime.modules.register.NewIdentityScreen
 import com.distributedLab.rarime.modules.rewards.RewardsClaimScreen
-import com.distributedLab.rarime.modules.rewards.RewardsEventItemScreen
+import com.distributedLab.rarime.modules.rewards.event_item.RewardsEventItemScreen
 import com.distributedLab.rarime.modules.rewards.RewardsScreen
 import com.distributedLab.rarime.modules.security.EnableBiometricsScreen
 import com.distributedLab.rarime.modules.security.EnablePasscodeScreen
